Doctor_Appointment_System/Doctor_Appointment_System_Project_Django/config/urls.py
"""config URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
import debug_toolbar
from django.contrib import admin
from django.urls import path
from django.urls import include
from userauth import views as ua_views

urlpatterns = [
    path('', include('appoint.urls')),
    path('profile/register', ua_views.register, name='register'),
    path('doctor/register', ua_views.register_doctor, name='register_doctor'),
    path('register', ua_views.register_moderator, name='register_moderator'),
    path('', include('django.contrib.auth.urls')),
    path('admin/', admin.site.urls),
    path('__debug__/', include(debug_toolbar.urls)),
]
