Doctor_Appointment_System/Doctor_Appointment_System_Project_Django/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-bootstrap4==3.0.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.1
psycopg2==2.8.6
PyGreSQL==5.2.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
